Added some calculation attribute rules related to work migrating Attribute Assistant (#72)
* Create CalcJunctionFields.md

First draft of this.  Code has been verified in Pro

* Update Attribute Assistant.md

Added links to Junction field Attribute Assistant rules.  The code in the CalcJunctionFields Arcade calculates values for both the from and to endpoints of the linear feature and attempts to find an intersecting point for each one and update fields for both points.

* Update CalcJunctionFields.md

Made it a little more generic

* Create CalcEdgeFields.md

Initial draft of this page

* Update Attribute Assistant.md

Added scripts for the From and To Edge rules

* Create SetMValues.md

Initial draft of the page.

* Update Attribute Assistant.md

Updated Set M Values rule.  Link to new code in the calculate rules folder.

* Updated links

Links made relative for the new rules

* Correct end_x/y evaluation, add ISEmptyButBetter

Logical fix.  2nd IF() statement needed to be checking the end values.
As suggested, used IsEmptyButBetter function.  Let me know if I'm using it correctly.

* Updated Comment

A reference above the code was left from a copy from another page.  Sorry about that.  The original statement holds.  This function works on a point layer intersecting with a linear layer.

* Added IsEmptyButBetter function

Revised based on suggestions about another script.

* Fixed If logic to "not"